rename speed and direction
Rename speed and direction columns and function names. Speed is used instead of meters_per_sec which was too specific since the result are CRS units per sec (except for EPSG:4326, where m/s is computed instead). Direction is used instead of heading since the result is the direction of movement and not the moving object's orientation